+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for yaml merge
 ### Fixed
 * DLT Debug mode when source is streaming and node is not
+* DataFrame type propagation when models are used as inputs to other models
 ### Updated
 * n/a
 ### Breaking changes